requirements.txt

accelerate==1.6.0
diffusers==0.35
transformers==4.50 
gradio==5.23.0
sentencepiece==0.2.0
pillow==11.1.0
av==12.1.0
numpy==1.26.2
scipy==1.12.0
requests==2.31.0
torchsde==0.2.6

einops
opencv-contrib-python
safetensors


qwen-vl-utils[decord]==0.0.8

protobuf==4.25.3